Reorganized nav_drawer and layouts main and login
Kingdom overview added
