+++ app/src/main/res/layout/activity_login.xml
@@ -1,25 +1,73 @@
 <?xml version="1.0" encoding="utf-8"?>
-    <android.support.v4.widget.DrawerLayout xmlns:android="http://schemas.android.com/apk/res/android"
-        xmlns:app="http://schemas.android.com/apk/res-auto"
-        xmlns:tools="http://schemas.android.com/tools"
-        android:id="@+id/drawer_layout"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        tools:openDrawer="start"
-        android:fitsSystemWindows="false">
+<RelativeLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent">
 
-               <include
-                    layout="@layout/app_bar_left_navigation"
-                    android:layout_width="match_parent"
-                    android:layout_height="match_parent" />
+    <EditText
+        android:id="@+id/loginName"
+        android:layout_width="fill_parent"
+        android:layout_height="wrap_content"
+        android:hint="Enter your login name"
+        android:layout_alignParentLeft="true"
+        android:layout_alignParentTop="true"
+        android:layout_alignParentStart="true"
+        android:layout_marginBottom="20dp"
+        android:layout_marginTop="20dp">
+    </EditText>
 
-        <android.support.design.widget.NavigationView
-            android:id="@+id/nav_view"
-            android:layout_width="wrap_content"
-            android:layout_height="match_parent"
-            android:layout_gravity="start"
-            android:fitsSystemWindows="true"
-            app:headerLayout="@layout/nav_header_left_navigation"
-            app:menu="@menu/activity_left_navigation_drawer" />
+    <EditText
+        android:id="@+id/loginPassword"
+        android:layout_width="fill_parent"
+        android:layout_height="wrap_content"
+        android:hint="Enter your login password"
+        android:layout_below="@+id/loginName"
+        android:layout_marginBottom="20dp"
+        android:layout_alignParentLeft="true"
+        android:inputType="textPassword">
+        <!--android:layout_alignParentStart="true"-->
+    </EditText>
 
-    </android.support.v4.widget.DrawerLayout>
+    <Button
+        android:onClick="saveData"
+        android:id="@+id/loginButton"
+        android:text="Login"
+        android:layout_below="@+id/loginPassword"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content">
+    </Button>
+
+    <Button
+        android:onClick="getData"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="Display Info"
+        android:id="@+id/displayButton"
+        android:layout_alignBottom="@+id/loginButton"
+        android:layout_alignParentRight="true"
+        android:layout_alignParentEnd="true">
+    </Button>
+
+    <TextView
+        android:text="Register Here"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:id="@+id/registerHereLink"
+        android:layout_below="@+id/loginButton"
+        android:layout_alignParentLeft="true"
+        android:layout_alignParentStart="true"
+        android:layout_marginTop="28dp"
+        android:layout_alignRight="@+id/loginButton"
+        android:layout_alignEnd="@+id/loginButton" />
+
+    <TextView
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:textAppearance="?android:attr/textAppearanceMedium"
+        android:id="@+id/dataView"
+        android:layout_below="@+id/loginButton"
+        android:layout_toLeftOf="@+id/displayButton"
+        android:layout_toStartOf="@+id/displayButton"
+        android:layout_marginTop="61dp" />
+
+</RelativeLayout>
