+++ app/src/main/res/layout/activity_main.xml
@@ -10,6 +10,31 @@
     android:paddingTop="@dimen/activity_vertical_margin"
     tools:context="com.greenfox.peridot.peridot_coz_android.activity.MainActivity">
 
+    <android.support.v4.widget.DrawerLayout xmlns:android="http://schemas.android.com/apk/res/android"
+        xmlns:app="http://schemas.android.com/apk/res-auto"
+        xmlns:tools="http://schemas.android.com/tools"
+        android:id="@+id/drawer_layout"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        tools:openDrawer="start"
+        android:fitsSystemWindows="false">
+
+        <include
+            layout="@layout/app_bar_left_navigation"
+            android:layout_width="match_parent"
+            android:layout_height="match_parent" />
+
+        <android.support.design.widget.NavigationView
+            android:id="@+id/nav_view"
+            android:layout_width="wrap_content"
+            android:layout_height="match_parent"
+            android:layout_gravity="start"
+            android:fitsSystemWindows="true"
+            app:headerLayout="@layout/nav_header_left_navigation"
+            app:menu="@menu/activity_left_navigation_drawer" />
+
+    </android.support.v4.widget.DrawerLayout>
+
     <TextView
         android:id="@+id/welcomeText"
         android:layout_width="wrap_content"
@@ -24,4 +49,5 @@
         android:layout_centerHorizontal="true"
         android:id="@+id/logoutButton"
         android:onClick="logout"/>
+
 </RelativeLayout>
