+++ app/src/main/res/layout/app_bar_left_navigation.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context="com.greenfox.peridot.peridot_coz_android.activity.LoginActivity">
+    tools:context="com.greenfox.peridot.peridot_coz_android.activity.MainActivity">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
@@ -27,4 +27,3 @@
 
 
 </android.support.design.widget.CoordinatorLayout>
-
