+++ app/src/main/res/layout/content_left_navigation.xml
@@ -10,7 +10,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="com.greenfox.peridot.peridot_coz_android.activity.LoginActivity"
+    tools:context="com.greenfox.peridot.peridot_coz_android.activity.MainActivity"
     tools:showIn="@layout/app_bar_left_navigation">
 
     <FrameLayout
@@ -18,77 +18,5 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent">
 
-    <RelativeLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent">
-
-        <EditText
-            android:id="@+id/loginName"
-            android:layout_width="fill_parent"
-            android:layout_height="wrap_content"
-            android:hint="Enter your login name"
-            android:layout_alignParentLeft="true"
-            android:layout_alignParentTop="true"
-            android:layout_alignParentStart="true"
-            android:layout_marginBottom="20dp"
-            android:layout_marginTop="20dp">
-        </EditText>
-
-        <EditText
-            android:id="@+id/loginPassword"
-            android:layout_width="fill_parent"
-            android:layout_height="wrap_content"
-            android:hint="Enter your login password"
-            android:layout_below="@+id/loginName"
-            android:layout_marginBottom="20dp"
-            android:layout_alignParentLeft="true"
-            android:inputType="textPassword">
-            <!--android:layout_alignParentStart="true"-->
-        </EditText>
-
-        <Button
-            android:onClick="saveData"
-            android:id="@+id/loginButton"
-            android:text="Login"
-            android:layout_below="@+id/loginPassword"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content">
-        </Button>
-
-        <Button
-            android:onClick="getData"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="Display Info"
-            android:id="@+id/displayButton"
-            android:layout_alignBottom="@+id/loginButton"
-            android:layout_alignParentRight="true"
-            android:layout_alignParentEnd="true">
-        </Button>
-
-        <TextView
-            android:text="Register Here"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:id="@+id/registerHereLink"
-            android:layout_below="@+id/loginButton"
-            android:layout_alignParentLeft="true"
-            android:layout_alignParentStart="true"
-            android:layout_marginTop="28dp"
-            android:layout_alignRight="@+id/loginButton"
-            android:layout_alignEnd="@+id/loginButton" />
-
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:textAppearance="?android:attr/textAppearanceMedium"
-            android:id="@+id/dataView"
-            android:layout_below="@+id/loginButton"
-            android:layout_toLeftOf="@+id/displayButton"
-            android:layout_toStartOf="@+id/displayButton"
-            android:layout_marginTop="61dp" />
-
-    </RelativeLayout>
     </FrameLayout>
-
 </RelativeLayout>
