+++ app/src/main/res/layout/kingdom_overview_layout.xml
@@ -4,11 +4,11 @@
     android:background="@android:color/holo_green_light">
 
     <TextView
-        android:text="You are on the first fragment layout"
+        android:text="This is your kingdom"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_alignParentTop="true"
         android:layout_centerHorizontal="true"
         android:layout_marginTop="151dp"
-        android:id="@+id/firstLayout" />
+        android:id="@+id/kingdomOverview" />
 </RelativeLayout>
