+++ app/src/main/res/menu/activity_left_navigation_drawer.xml
@@ -3,9 +3,9 @@
 
     <group android:checkableBehavior="single">
         <item
-            android:id="@+id/nav_first_layout"
+            android:id="@+id/nav_kingdom_overview"
             android:icon="@drawable/ic_menu_camera"
-            android:title="First Layout" />
+            android:title="Your Kingdom" />
         <item
             android:id="@+id/nav_second_layout"
             android:icon="@drawable/ic_menu_gallery"
